+++ Task_5/Models/UserSettings.cs
@@ -4,7 +4,7 @@
     {
         public string region { get; set; } = "ru";
         public int seed { get; set; } = 0;
-        public int errorValue {  get; set; } = 0;
+        public double errorValue {  get; set; } = 0;
 
     }
 }
